Major rewrite of issue processing.
As it turns out, a GitHub repo's pull requests share an ID namespace with
the issue tracker. So our assumption that we could enforce a mapping between
Google Code issue IDs and GitHub IDs was dead wrong.

This change rewrites the issue exporting script to simply export issues ignoring
the Google Code issue IDs. This simplifies the code considerably.